feat: TTS 프로필 삭제

+++ backend/src/main/java/com/checkitout/ijoa/TTS/repository/TTSRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TTSRepository extends JpaRepository<TTS, Integer> {
+public interface TTSRepository extends JpaRepository<TTS, Long> {
 }
